requirements.txt
altair==6.0.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
cachetools==7.0.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
distro==1.9.0
docx==0.2.4
docx2pdf==0.1.8
filelock==3.20.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.2
mercadopago==2.3.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.17.0
networkx==3.6.1
numba==0.64.0
numpy==1.26.4
openai-whisper==20250625
packaging==26.0
pandas==2.3.3
pefile==2024.8.26
pillow==12.0.0
platformdirs==4.9.2
pooch==1.9.0
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.2
pyparsing==3.3.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
pywin32==311
pywin32-ctypes==0.2.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
wheel==0.46.3
